Add files via upload
try to get the music working, there are time stamps of part of the song. pick one then we can try settimeout on classes if not we just dump them in the main code.
team, one has 54 second of action to do 

In the first 16 second it will be a slow build up maybe snow effect with a simple straight bar line with black background. 

The 20 second would be a blue heat appearing in the middle of the screen beating with the song 

Add whatever you want. 

53 second snow full the screen. 

The second team code will happen on white background. 

Team two have 64 second of action all of it is high energy mainly use sharp shapes 

 

Team three is techno so do random colours RGB. 

Use round shape